Create sleepy-cow-herding.cpp
Almost first try. Genuinely was so easy and am lokey annoyed as I got it wrong the first time as the problem insinuated the cows were not consecutive and had to be moved. So, I jus thad to add one more if statement for the one test case that had three consecutive numbers.